backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]",
    "fastapi>=0.122.1",
    "passlib[argon2]>=1.7.4",
    "pydantic[email]>=2.12.5",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.44",
    "pyjwt>=2.10.1",
    "pydantic-settings>=2.12.0",
]
